requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.3
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.1
django-celery-beat==2.4.0
django-choices==1.7.2
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django-mathfilters==1.0.0
django-modeltranslation==0.18.7
django-rosetta==0.9.8
django-timezone-field==5.0
django-utils-six==2.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
filelock==3.8.0
gunicorn==20.1.0
idna==3.4
kombu==5.2.4
Markdown==3.4.1
oauthlib==3.2.2
packaging==21.3
Pillow==9.2.0
platformdirs==2.5.3
pluggy==1.0.0
polib==1.1.1
prompt-toolkit==3.0.36
psycopg2-binary==2.9.4
py==1.11.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.6
redis==4.4.2
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
stripe==5.0.0
text-unidecode==1.3
tomli==2.0.1
tox==3.27.0
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.12
vine==5.0.0
virtualenv==20.16.6
wcwidth==0.2.6
whitenoise==6.3.0
